factor out dependency graph from to_anf pass for pretty printer
