+++ src/relay/pass/to_a_normal_form.cc
@@ -29,192 +29,10 @@
 #include "let_list.h"
 #include "../../common/arena.h"
 #include "pass_util.h"
+#include "dependency_graph.h"
 
 namespace tvm {
 namespace relay {
-
-using common::LinkNode;
-using common::LinkedList;
-
-/* DependencyGraph track input and output of an Expr.
- * Additionally, dummy scope is created to model scope.
- * It allow us to traverse the graph in reverse order.
- */
-class DependencyGraph {
- public:
-  /*! \brief A node in the graph. */
-  struct Node {
-    bool new_scope = false;
-    LinkedList<Node*> input;
-    LinkedList<Node*> output;
-  };
-
-  /*! \brief The node map that maps node to graph */
-  std::unordered_map<Expr, Node*, NodeHash, NodeEqual> expr_node;
-
-  /*! \brief All the nodes in post DFS order */
-  std::vector<Node*> post_dfs_order;
-
-  /*!
-   * \brief create a dependency graph.
-   * \param arena The arena used for data allocation.
-   * \param body The body of the expression to create a graph.
-   */
-  static DependencyGraph Create(common::Arena* arena, const Expr& body);
-
- private:
-  class Creator;
-};
-
-// Creator of DependencyGraph
-class DependencyGraph::Creator : private ExprFunctor<void(const Expr& e)> {
- public:
-  explicit Creator(common::Arena* arena)
-    : arena_(arena) {}
-
-  DependencyGraph Create(const Expr& body) {
-    this->VisitExpr(body);
-    return std::move(graph_);
-  }
-
- private:
-  /*! \brief allocator of all the internal node object */
-  common::Arena* arena_;
-  // The output.
-  DependencyGraph graph_;
-  // Update the message stored at the node.
-  void Depend(DependencyGraph::Node* parent, const Expr& child) {
-    VisitExpr(child);
-
-    CHECK_NE(graph_.expr_node.count(child), 0);
-
-    Depend(parent, graph_.expr_node[child]);
-  }
-
-  void Depend(DependencyGraph::Node* parent, DependencyGraph::Node* child) {
-    auto* parent_link = arena_->make<LinkNode<DependencyGraph::Node*> >();
-    parent_link->value = parent;
-    child->output.Push(parent_link);
-
-    auto* child_link = arena_->make<LinkNode<DependencyGraph::Node*> >();
-    child_link->value = child;
-    parent->input.Push(child_link);
-  }
-
-  std::unordered_set<Expr, NodeHash, NodeEqual> visited_;
-
-  DependencyGraph::Node* NewNode(bool new_scope) {
-    auto* ret = arena_->make<DependencyGraph::Node>();
-    ret->new_scope = new_scope;
-    return ret;
-  }
-
-  void VisitExpr(const Expr& e) final {
-    if (visited_.count(e) == 0) {
-      if (graph_.expr_node.count(e) == 0) {
-        graph_.expr_node[e] = NewNode(false);
-      }
-      visited_.insert(e);
-      ExprFunctor<void(const Expr&)>::VisitExpr(e);
-      graph_.post_dfs_order.push_back(graph_.expr_node[e]);
-    }
-  }
-
-  void VisitExpr_(const CallNode* c) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(c)];
-    Depend(n, c->op);
-    for (const auto& a : c->args) {
-      Depend(n, a);
-    }
-  }
-
-  void VisitExpr_(const TupleNode* t) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(t)];
-    for (const auto& a : t->fields) {
-      Depend(n, a);
-    }
-  }
-
-  void VisitExpr_(const TupleGetItemNode* t) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(t)];
-    Depend(n, t->tuple);
-  }
-
-  void VisitExpr_(const RefCreateNode* r) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(r)];
-    Depend(n, r->value);
-  }
-
-  void VisitExpr_(const RefReadNode* r) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(r)];
-    Depend(n, r->ref);
-  }
-
-  void VisitExpr_(const RefWriteNode* r) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(r)];
-    Depend(n, r->ref);
-    Depend(n, r->value);
-  }
-
-  void VisitExpr_(const IfNode* i) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(i)];
-    DependencyGraph::Node* t = NewNode(true);
-    DependencyGraph::Node* f = NewNode(true);
-    Depend(n, i->cond);
-    Depend(n, t);
-    Depend(n, f);
-    Depend(t, i->true_branch);
-    Depend(f, i->false_branch);
-    graph_.post_dfs_order.push_back(f);
-    graph_.post_dfs_order.push_back(t);
-  }
-
-  void VisitExpr_(const FunctionNode* f) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(f)];
-    DependencyGraph::Node* b = NewNode(true);
-    Depend(n, b);
-    Depend(b, f->body);
-    graph_.post_dfs_order.push_back(b);
-  }
-
-  void VisitExpr_(const LetNode* l) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(l)];
-    DependencyGraph::Node* b = NewNode(true);
-    Depend(n, b);
-    Depend(b, l->value);
-    Depend(b, l->body);
-    graph_.post_dfs_order.push_back(b);
-  }
-
-  void VisitExpr_(const MatchNode* m) final {
-    DependencyGraph::Node* n = graph_.expr_node[GetRef<Expr>(m)];
-    Depend(n, m->data);
-    std::vector<DependencyGraph::Node*> v;
-    for (const Clause& c : m->clauses) {
-      DependencyGraph::Node* b = NewNode(true);
-      Depend(n, b);
-      Depend(b, c->rhs);
-      v.push_back(b);
-    }
-    for (auto it = v.rbegin(); it != v.rend(); ++it) {
-      graph_.post_dfs_order.push_back(*it);
-    }
-  }
-
-  void VisitExpr_(const VarNode* v) final { }
-
-  void VisitExpr_(const GlobalVarNode* v) final { }
-
-  void VisitExpr_(const ConstantNode* c) final { }
-
-  void VisitExpr_(const OpNode* o) final { }
-
-  void VisitExpr_(const ConstructorNode* c) final { }
-};
-
-DependencyGraph DependencyGraph::Create(common::Arena* arena, const Expr& body) {
-  return Creator(arena).Create(body);
-}
 
 Expr ToANormalForm(const Expr& e, const Module& m, std::set<GlobalVar>* gv);
 
@@ -256,7 +74,7 @@ std::unordered_map<DependencyGraph::Node*, Scope> CalcScope(const DependencyGrap
   Scope global_scope = std::make_shared<ScopeNode>();
   for (auto it = dg.post_dfs_order.rbegin(); it != dg.post_dfs_order.rend(); ++it) {
     DependencyGraph::Node* n = *it;
-    auto iit = n->output.head;
+    auto iit = n->parents.head;
     Scope s;
     if (iit == nullptr) {
       s = global_scope;
@@ -313,7 +131,7 @@ class Fill : ExprFunctor<Expr(const Expr&, const Var&)> {
 
   Scope GetSubScope(const Expr& e, size_t i) {
     DependencyGraph::Node* n = dg_.expr_node.at(e);
-    auto h = n->input.head;
+    auto h = n->children.head;
     while (i != 0) {
       CHECK(h);
       --i;
